+++ Sisfarma.Sincronizador.Domain.Entities/Fisiotes/Lista.cs
@@ -2,7 +2,7 @@ namespace Sisfarma.Sincronizador.Domain.Entities.Fisiotes
 {
     public partial class Lista
     {
-        public int cod { get; set; }
+        public long cod { get; set; }
 
         public string lista { get; set; }
